optimize invokeinterface calls with hash-based interface method tables
Previously, the JIT compiler implemented invokeinterface by calling a
helper thunk to find the appropriate method and then calling the
result, which was slow because it involved an out-of-line call and a
linear search over all the implemented interfaces for the instance.

This commit improves upon that strategy by adding a fixed-size area to
each class's vtable for interface calls.  This area is initialized by
hashing the name and spec of each interface method implemented and
placing the address of a thunk at that location.  The thunk to use
depends on whether there is a collision (more than one method maps to
that index into the table) and thus whether we should later overwrite
the address of the thunk with the address of the compiled method.  In
the common case of a unique mapping (i.e. no collision), the result is
a call which is almost as fast as an invokevirtual call.  In the case
of a collision, we revert to the old algorithm.

This approach was inspired by this abstract:

http://www.research.ibm.com/people/d/dgrove/papers/oopsla01.html

The new strategy involves a few trade-offs:

 * We reserve space for the interface method table for all classes,
   not just those which implement non-empty interfaces.  We could
   avoid this by allocating it separately instead of making it part of
   the vtable, but this would add another level of indirection to each
   interface call.

 * Classes which implement only a few interface methods end up with a
   sparsely-populated IMT, which wastes memory.  We can mitigate this
   my shrinking the IMT, but that will increase the number of
   collisions.

 * Even in the case of a unique mapping, invokeinterface calls still
   aren't quite as fast as invokevirtual calls since we still have to
   store a reference to the actual method to be called in a known
   location, since we can't be sure when generating the call that
   there won't be a collision or that the callee has already been
   compiled.  One way around this is to maintain a map of call sites
   to target methods and use that to derive the target from the thunk.
   However, that only works if the thunk has a return address that
   points to the call site, which isn't always the case for tail
   calls.  Another thing we can do is make the store operation as
   inexpensive as possible by loading to a known register instead of a
   thread-local field, hence a todo comment to that effect in the
   code.

This commit also includes some peripheral cleanups involving bootstrap
classes needed to accomodate the new strategy.