optimize invokeinterface for TailCalls=false case
When we're not optimizing tail calls by reusing stack frames, we can
rely on the return address always pointing to the call site, which
means we need not generate code to store a reference to the target
method in a known location when compiling invokeinterface.  Instead,
we add a mapping from the call site address to the target method in
MyProcessor::callTable when compiling the caller, which we can later
refer to from compileInterfaceMethod2 to determine the appropriate
method to look up.