I've released Avian-pack
This is a small addition to Avian page with a brief description for Avian-pack and a link to the release page.

May I ask you to add it? I want some user attention for inspiration :)
If you find some grammar mistakes in the text, please, let me know, cause my English is bad.

Thank you!